refactor: Create util package
